stop using bytes for indexing

+++ source/JSON/JSONArray.cs
@@ -51,7 +51,7 @@ namespace Serialization.JSON
             if (value->elements.Count > 0)
             {
                 NewLine();
-                for (byte i = 0; i <= depth; i++)
+                for (int i = 0; i <= depth; i++)
                 {
                     Indent(indent);
                 }
@@ -72,14 +72,14 @@ namespace Serialization.JSON
 
                     result.Append(',');
                     NewLine();
-                    for (byte i = 0; i <= depth; i++)
+                    for (int i = 0; i <= depth; i++)
                     {
                         Indent(indent);
                     }
                 }
 
                 NewLine();
-                for (byte i = 0; i < depth; i++)
+                for (int i = 0; i < depth; i++)
                 {
                     Indent(indent);
                 }
